pyproject.toml
[tool.poetry]
name = "tstickers"
version = "2024"
license = "mit"
description = "Download sticker packs from Telegram"
authors = ["FredHappyface"]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities",
	"Topic :: Multimedia :: Graphics",
]
homepage = "https://github.com/FHPythonUtils/TStickers"
repository = "https://github.com/FHPythonUtils/TStickers"
documentation = "https://github.com/FHPythonUtils/TStickers/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "<3,>=2.31.0"
Pillow = "<11,>=10.2.0"
emoji = "<3,>=2.10.1"
requests-cache = "<2,>=1.2.0"
rlottie-python = { version = "<2,>=1.3.3", optional = true }
pyrlottie = { version = "<2026,>=2024.0.1", optional = true }
loguru = "<2,>=0.7.2"

[tool.poetry.scripts]
tstickers = "tstickers:cli"

[tool.poetry.extras]
rlottie-python = ["rlottie-python"]
pyrlottie = ["pyrlottie"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
handsdown = "^2.1.0"
coverage = "^7.4.4"
ruff = "^0.2.2"
pyright = "^1.1.354"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
	"ANN101",  # type annotation for self in method
	"COM812",  # enforce trailing comma
	"D2",  # pydocstyle formatting
	"ISC001",
	"N",  # pep8 naming
	"PLR09",  # pylint refactor too many
	"TCH",  # type check blocks
	"W191"  # ignore this to allow tabs
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["D", "S101", "E402"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all" # Disallow all relative imports.

[tool.ruff.format]
indent-style = "tab"
docstring-code-format = true
line-ending = "lf"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.coverage.run]
branch = true

[tool.tox]
legacy_tox_ini = """
[tox]
env_list =
	py311
	py310
	py39
	py38

[testenv]
deps =
	imgcompare
	pytest
commands = pytest tests
"""
